requirements.txt
setuptools
numpy<2.0.0
opencv-python==4.8.1.78
Pillow==10.0.0
pytesseract==0.3.10
PyPDF2==3.0.1
